Split out helper classes and static methods from Reader
