+++ src/CMakeLists.txt
@@ -26,6 +26,7 @@ add_library(mpims SHARED
   mpims.cc
   MSColumns.cc
   MSArray.cc
+  ReaderBase.cc
   Reader.cc
   Writer.cc)
 target_include_directories(mpims PRIVATE
@@ -48,14 +49,18 @@ install(TARGETS mpims
   GROUP_READ GROUP_EXECUTE WORLD_READ WORLD_EXECUTE)
 install(FILES
   ArrayIndexer.h
+  AxisIter.h
   ColumnAxis.h
   DataDistribution.h
   IndexBlock.h
+  IterParams.h
   mpims.h
   MSArray.h
   MSColumns.h
+  ReaderBase.h
   Reader.h
   ReaderMPIState.h
+  TraversalState.h
   Writer.h
   DESTINATION include
   CONFIGURATIONS Release
