+++ src/Reader.h
@@ -24,6 +24,8 @@
 #include <MSColumns.h>
 #include <DataDistribution.h>
 #include <MPIState.h>
+#include <TraversalState.h>
+#include <ReaderBase.h>
 
 namespace mpims {
 
@@ -32,303 +34,20 @@ enum AMode {
   ReadWrite
 };
 
+class IterParams;
+class AxisIter;
+
 class Reader
-  : public std::iterator<std::input_iterator_tag, const MSArray, std::size_t> {
+  : public ReaderBase
+  , public std::iterator<std::input_iterator_tag, const MSArray, std::size_t> {
 
   friend class Writer;
 
 public:
 
-  // Array with an indeterminate, internal, that is, any but the outermost, axis
-  class IndeterminateArrayError
-    : std::runtime_error {
-  public:
-    explicit IndeterminateArrayError()
-      : std::runtime_error("indeterminate internal array axis") {
-    }
-  };
-
-  // Traversal of an indeterminate MS with the top traversal axis not equal to
-  // the top (indeterminate) MS axis.
-  class IndeterminateArrayTraversalError
-    : std::runtime_error {
-  public:
-    explicit IndeterminateArrayTraversalError()
-      : std::runtime_error(
-        "top of traversal order not the indeterminate axis") {
-    }
-  };
-
-  class IterParams;
-  class AxisIter;
-
-  // TraversalState maintains the state used to traverse the MS in the order
-  // provided by the client. If one considers MS traversal as stack of
-  // traversals along every axis with recursion to get to the next axis, this
-  // structure contains the information needed for traversal after flattening
-  // such a recursive iteration into a single loop.
-  struct TraversalState {
-
-    TraversalState()
-      : cont(false)
-      , global_eof(true) {
-    }
-
-    TraversalState(
-      const std::shared_ptr<const std::vector<IterParams> >& iter_params_,
-      MSColumns outer_ms_axis,
-      std::size_t outer_ms_length,
-      bool in_tail_,
-      MSColumns outer_full_array_axis_,
-      const std::shared_ptr<const MPI_Datatype>& full_buffer_datatype,
-      unsigned full_buffer_dt_count,
-      const std::shared_ptr<const MPI_Datatype>& tail_buffer_datatype,
-      unsigned tail_buffer_dt_count,
-      const std::shared_ptr<const MPI_Datatype>& full_fileview_datatype,
-      const std::shared_ptr<const MPI_Datatype>& tail_fileview_datatype)
-    : iter_params(iter_params_)
-    , block_maps(make_index_block_sequences(iter_params_))
-    , count(0)
-    , max_count(0)
-    , in_tail(in_tail_)
-    , outer_full_array_axis(outer_full_array_axis_)
-    , m_full_buffer_datatype(full_buffer_datatype)
-    , m_full_buffer_dt_count(full_buffer_dt_count)
-    , m_tail_buffer_datatype(tail_buffer_datatype)
-    , m_tail_buffer_dt_count(tail_buffer_dt_count)
-    , m_full_fileview_datatype(full_fileview_datatype)
-    , m_tail_fileview_datatype(tail_fileview_datatype) {
-
-      global_eof = false;
-      const IterParams* init_params = &(*iter_params)[0];
-      axis_iters.emplace_back(
-        std::shared_ptr<const IterParams>(iter_params, init_params),
-        !init_params->max_blocks || init_params->max_blocks > 0);
-      std::for_each(
-        std::begin(*iter_params),
-        std::end(*iter_params),
-        [this](const IterParams& ip) {
-          data_index[ip.axis] = ip.origin;
-        });
-
-      if (axis_iters.front().params->axis == outer_ms_axis) {
-        eof_axis_iters.push_back(axis_iters.front());
-        auto& eai = eof_axis_iters.front();
-        max_count = std::max(eai.params->buffer_capacity, 1uL);
-        while (!eai.at_end && eai.index < outer_ms_length)
-          eai.increment();
-      }
-      cont = true;
-    }
-
-    //EOF condition flag
-    bool
-    eof() const {
-      if (axis_iters.empty() || global_eof)
-        return true;
-      if (eof_axis_iters.empty())
-        return false;
-      auto ai = std::rbegin(axis_iters);
-      auto ai_end = std::rend(axis_iters);
-      auto eai = std::rbegin(eof_axis_iters);
-      auto eai_end = std::rend(eof_axis_iters);
-      bool result = false;
-      auto num_undefined = axis_iters.size() - eof_axis_iters.size();
-      while (num_undefined-- > 0 && ai != ai_end)
-        ++ai;
-      while (!result && ai != ai_end && eai != eai_end) {
-        result = ai->index >= eai->index;
-        ++ai;
-        ++eai;
-      }
-      return result;
-    };
-
-    bool
-    at_end() const {
-      return !cont || global_eof;
-    }
-
-    // continuation condition flag (to allow breaking out from iteration)
-    bool cont;
-
-    bool global_eof;
-
-    std::shared_ptr<const std::vector<IterParams> > iter_params;
-
-    std::shared_ptr<
-      const std::vector<IndexBlockSequenceMap<MSColumns> > > block_maps;
-
-    ArrayIndexer<MSColumns>::index data_index;
-
-    // stack of AxisIters to maintain axis iteration indexes
-    std::deque<AxisIter> axis_iters;
-
-    std::deque<AxisIter> eof_axis_iters;
-
-    int count;
-
-    int max_count;
-
-    // Is traversal currently in a tail fileview? A tail fileview differs
-    // somewhat from a non-tail fileview when the iteration at the outermost
-    // fileview axis doesn't fit into the same pattern at the end of the axis as
-    // all the previous fileviews. This can happen when the distribution of data
-    // on an axis to processes is uneven at the end of an axis.
-    bool in_tail;
-
-    // the outermost axis at which data can fully fit into a buffer
-    MSColumns outer_full_array_axis;
-
-    std::tuple<std::shared_ptr<const MPI_Datatype>, unsigned>
-    buffer_datatype() const {
-      if (in_tail)
-        return std::make_tuple(m_tail_buffer_datatype, m_tail_buffer_dt_count);
-      else
-        return std::make_tuple(m_full_buffer_datatype, m_full_buffer_dt_count);
-    }
-
-    std::shared_ptr<const MPI_Datatype>
-    fileview_datatype() const {
-      return (in_tail ? m_tail_fileview_datatype : m_full_fileview_datatype);
-    }
-
-    bool
-    operator==(const TraversalState& other) const {
-      return (
-        cont == other.cont
-        && count == other.count
-        && (block_maps == other.block_maps || *block_maps == *other.block_maps)
-        && data_index == other.data_index
-        && axis_iters == other.axis_iters
-        && eof_axis_iters == other.eof_axis_iters
-        && in_tail == other.in_tail
-        && m_tail_buffer_dt_count == other.m_tail_buffer_dt_count
-        && m_full_buffer_dt_count == other.m_full_buffer_dt_count
-        // comparing array datatypes and fileview datatypes would be ideal, but
-        // that could be a relatively expensive task, as it requires digging
-        // into the contents of the datatypes; instead, since those datatypes
-        // are functions of the MS shape, iteration order and outer array axis
-        // alone, given the previous comparisons, we can just compare the outer
-        // arrays axis values
-        && outer_full_array_axis == other.outer_full_array_axis);
-    }
-
-    bool
-    operator!=(const TraversalState& other) const {
-      return !operator==(other);
-    }
-
-    std::vector<IndexBlockSequence<MSColumns> >
-    blocks(int cnt) const {
-      std::vector<IndexBlockSequence<MSColumns> > result;
-      if (cnt > 0) {
-        for (std::size_t i = 0; i < iter_params->size(); ++i) {
-          auto& ip = (*iter_params)[i];
-          if (!ip.fully_in_array && ip.buffer_capacity == 0) {
-            result.emplace_back(
-              ip.axis,
-              std::vector{ IndexBlock(data_index.at(ip.axis), 1) });
-          } else {
-            auto& map = (*block_maps)[i];
-            auto ibs = map[data_index.at(ip.axis)];
-            ibs.trim();
-            result.push_back(ibs);
-          }
-        }
-      }
-      return result;
-    }
-
-    void
-    advance_to_buffer_end() {
-      if (!axis_iters.empty()) {
-        AxisIter* axis_iter = &axis_iters.back();
-        axis_iter->increment(max_count);
-        data_index[axis_iter->params->axis] = axis_iter->index;
-        while (!axis_iters.empty() && axis_iter->at_end) {
-          data_index[axis_iter->params->axis] = axis_iter->params->origin;
-          axis_iters.pop_back();
-          if (!axis_iters.empty()) {
-            axis_iter = &axis_iters.back();
-            axis_iter->increment();
-            data_index[axis_iter->params->axis] = axis_iter->index;
-          }
-        }}
-    }
-
-    template <typename F>
-    void
-    advance_to_next_buffer(
-      std::shared_ptr<const std::optional<MSColumns> > inner_fileview_axis,
-      F at_fileview_axis) {
-
-      count = 0;
-      max_count = 0;
-      while (!axis_iters.empty()) {
-        AxisIter& axis_iter = axis_iters.back();
-        MSColumns axis = axis_iter.params->axis;
-        if (!axis_iter.at_end) {
-          auto depth = axis_iters.size();
-          data_index[axis] = axis_iter.index;
-          if (axis_iter.params->buffer_capacity > 0) {
-            max_count = static_cast<int>(axis_iter.params->buffer_capacity);
-            auto nr = axis_iter.num_remaining();
-            if (nr && (nr.value() < static_cast<std::size_t>(max_count))) {
-              count = nr.value();
-              in_tail = true;
-            } else {
-              count = max_count;
-              in_tail = false;
-            }
-          } else {
-            in_tail = false;
-          }
-          if (*inner_fileview_axis && axis == inner_fileview_axis->value())
-            at_fileview_axis(axis);
-          if (axis_iter.params->buffer_capacity > 0)
-            return;
-          const IterParams* next_params = &(*iter_params)[depth];
-          axis_iters.emplace_back(
-            std::shared_ptr<const IterParams>(iter_params, next_params),
-            axis_iter.at_data);
-        } else {
-          data_index[axis] = axis_iter.params->origin;
-          axis_iters.pop_back();
-          if (!axis_iters.empty())
-            axis_iters.back().increment();
-        }
-      }
-    }
-
-  private:
-
-    std::shared_ptr<const MPI_Datatype> m_full_buffer_datatype;
-
-    unsigned m_full_buffer_dt_count;
-
-    std::shared_ptr<const MPI_Datatype> m_tail_buffer_datatype;
-
-    unsigned m_tail_buffer_dt_count;
-
-    std::shared_ptr<const MPI_Datatype> m_full_fileview_datatype;
-
-    std::shared_ptr<const MPI_Datatype> m_tail_fileview_datatype;
-
-    std::ostringstream
-    show_iters(const std::deque<AxisIter>& iters) const {
-      std::ostringstream result;
-      std::for_each(
-        std::begin(iters),
-        std::end(iters),
-        [&result](const auto& ai) {
-          result << mscol_nickname(ai.params->axis) << ":" << ai.index << ";";
-        });
-      return result;
-    }
-  };
-
-  Reader();
+  Reader()
+    : m_readahead(false) {
+  }
 
   Reader(
     MPIState&& mpi_state,
@@ -418,150 +137,6 @@ public:
   }
 
   ~Reader();
-
-  // IterParams holds the information for the iteration over a single axis, as
-  // well as context within the scope of iteration over the entire MS together
-  // given a user buffer size. Most of the values in this structure are used to
-  // describe the iteration over an axis for a particular process.
-  struct IterParams {
-    MSColumns axis;
-    // can a buffer hold the data for the full axis given iteration pattern?
-    bool fully_in_array;
-    // is iteration across this axis done without changing the fileview?
-    bool within_fileview;
-    // number of axis values in iteration for which the entire array comprising
-    // data from all deeper axes can be copied into a single buffer
-    std::size_t buffer_capacity;
-    // number of data values in array comprising data from all deeper axes
-    std::size_t array_length;
-    // remainder are values describing iteration pattern
-    std::size_t origin, stride, block_len,
-      terminal_block_len, max_terminal_block_len;
-    std::optional<std::size_t> length, max_blocks;
-
-    bool
-    operator==(const IterParams& rhs) const {
-      return (
-        axis == rhs.axis
-        && fully_in_array == rhs.fully_in_array
-        && within_fileview == rhs.within_fileview
-        && buffer_capacity == rhs.buffer_capacity
-        && length == rhs.length
-        && origin == rhs.origin
-        && stride == rhs.stride
-        && block_len == rhs.block_len
-        && max_blocks == rhs.max_blocks
-        && terminal_block_len == rhs.terminal_block_len
-        && max_terminal_block_len == rhs.max_terminal_block_len);
-    }
-
-    bool
-    operator!=(const IterParams& rhs) const {
-      return !(operator==(rhs));
-    }
-
-    std::optional<std::size_t>
-    accessible_length() const {
-      if (max_blocks.has_value())
-        return block_len * (max_blocks.value() - 1) + terminal_block_len;
-      else
-        return std::nullopt;
-    }
-
-    std::optional<std::size_t>
-    max_accessible_length() const {
-      if (max_blocks.has_value())
-        return block_len * (max_blocks.value() - 1) + max_terminal_block_len;
-      else
-        return std::nullopt;
-    }
-  };
-
-  // AxisIter contains the state of the iteration along an axis. It follows the
-  // iteration pattern defined for an axis by an IterParams value.
-  class AxisIter {
-  public:
-
-    AxisIter(std::shared_ptr<const IterParams> params_, bool outer_at_data_)
-      : params(params_)
-      , index(params_->origin)
-      , block(0)
-      , at_data((!params_->max_blocks || params_->max_blocks.value() > 0)
-                && outer_at_data_)
-      , outer_at_data(outer_at_data_)
-      , at_end(params_->max_blocks && params_->max_blocks.value() == 0) {
-    }
-
-    bool
-    operator==(const AxisIter& rhs) const {
-      return *params == *rhs.params && index == rhs.index && block == rhs.block
-        && at_data == rhs.at_data && outer_at_data == rhs.outer_at_data
-        && at_end == rhs.at_end;
-    }
-
-    bool
-    operator!=(const AxisIter& rhs) const {
-      return !(operator==(rhs));
-    }
-
-    std::shared_ptr<const IterParams> params;
-    std::size_t index;
-    std::size_t block;
-    bool at_data;
-    bool outer_at_data;
-    bool at_end;
-
-    void
-    increment(std::size_t n=1) {
-      while (n > 0 && !at_end) {
-        block = index / params->stride;
-        auto block_origin = params->origin + block * params->stride;
-        bool terminal_block =
-          (params->max_blocks
-           ? (block == params->max_blocks.value() - 1)
-           : false);
-        auto block_len =
-          (terminal_block ? params->terminal_block_len : params->block_len);
-        ++index;
-        if (!terminal_block && index - block_origin >= block_len) {
-          ++block;
-          index = params->origin + block * params->stride;
-          block_origin = index;
-          terminal_block = (
-            params->max_blocks
-            ? (block == params->max_blocks.value() - 1)
-            : false);
-          block_len =
-            (terminal_block ? params->terminal_block_len : params->block_len);
-        }
-        at_data = outer_at_data && (index - block_origin < block_len);
-        at_end =
-          terminal_block
-          && (index - block_origin == params->max_terminal_block_len);
-        --n;
-      }
-    }
-
-    std::optional<std::size_t>
-    num_remaining() const {
-      if (at_end || !at_data)
-        return 0;
-      auto accessible_length = params->accessible_length();
-      if (accessible_length) {
-        auto block_origin = params->origin + block * params->stride;
-        return (accessible_length.value()
-                - (block * params->block_len + index - block_origin));
-      } else {
-        return std::nullopt;
-      }
-    }
-
-    void
-    complete() {
-      at_data = false;
-      at_end = true;
-    }
-  };
 
   static Reader
   begin(
@@ -727,24 +302,6 @@ protected:
     return m_ms_array;
   };
 
-  static void
-  init_iterparams(
-    const std::vector<ColumnAxisBase<MSColumns> >& ms_shape,
-    const std::vector<MSColumns>& traversal_order,
-    const std::unordered_map<MSColumns, DataDistribution>& pgrid,
-    int rank,
-    bool debug_log,
-    std::shared_ptr<std::vector<IterParams> >& iter_params);
-
-  static void
-  init_traversal_partitions(
-    const std::vector<ColumnAxisBase<MSColumns> >& ms_shape,
-    std::size_t& buffer_size,
-    std::shared_ptr<std::vector<IterParams> >& iter_params,
-    std::shared_ptr<std::optional<MSColumns> >& inner_fileview_axis,
-    int rank,
-    bool debug_log);
-
   static std::tuple<std::unique_ptr<MPI_Datatype, DatatypeDeleter>, unsigned>
   init_buffer_datatype(
     const std::vector<ColumnAxisBase<MSColumns> >& ms_shape,
@@ -753,42 +310,6 @@ protected:
     bool tail_array,
     int rank,
     bool debug_log);
-
-  static std::tuple<
-    std::unique_ptr<MPI_Datatype, DatatypeDeleter>,
-    std::size_t>
-  vector_datatype(
-    MPI_Aint value_extent,
-    std::unique_ptr<MPI_Datatype, DatatypeDeleter>& dt,
-    std::size_t dt_extent,
-    std::size_t offset,
-    std::size_t num_blocks,
-    std::size_t block_len,
-    std::size_t terminal_block_len,
-    std::size_t stride,
-    std::size_t len,
-    int rank,
-    bool debug_log);
-
-  static std::tuple<
-    std::unique_ptr<MPI_Datatype, DatatypeDeleter>,
-    std::size_t,
-    bool>
-  compound_datatype(
-    MPI_Aint value_extent,
-    std::unique_ptr<MPI_Datatype, DatatypeDeleter>& dt,
-    std::size_t dt_extent,
-    std::size_t offset,
-    std::size_t stride,
-    std::size_t num_blocks,
-    std::size_t block_len,
-    std::size_t terminal_block_len,
-    const std::optional<std::size_t>& len,
-    int rank,
-    bool debug_log);
-
-  static std::tuple<std::optional<std::tuple<std::size_t, std::size_t> >, bool>
-  tail_buffer_blocks(const IterParams& ip);
 
   static std::unique_ptr<MPI_Datatype, DatatypeDeleter>
   init_fileview(
@@ -799,10 +320,6 @@ protected:
     bool tail_fileview,
     int rank,
     bool debug_log);
-
-  static std::unique_ptr<std::vector<IndexBlockSequenceMap<MSColumns> > >
-  make_index_block_sequences(
-    const std::shared_ptr<const std::vector<IterParams> >& iter_params);
 
   MSArray
   read_arrays(
@@ -835,23 +352,6 @@ protected:
 
   bool
   buffer_order_compare(const MSColumns& col0, const MSColumns& col1) const;
-
-  static const IterParams*
-  find_iter_params(
-    const std::shared_ptr<const std::vector<IterParams> >& iter_params,
-    MSColumns col) {
-
-    auto ip =
-      std::find_if(
-        std::begin(*iter_params),
-        std::end(*iter_params),
-        [&col](auto& ip) {
-          return ip.axis == col;
-        });
-    if (ip != std::end(*iter_params))
-      return &*ip;
-    return nullptr;
-  }
 
   std::vector<IndexBlockSequence<MSColumns> >
   sorted_blocks(const TraversalState& state) const;
